+++ tileops/ops/attention/deepseek_dsa_decode.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.attention import SparseMlaKernel
 from tileops.kernels.kernel_base import Kernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["DeepSeekSparseAttentionDecodeWithKVCacheFwdOp"]
