+++ tileops/ops/attention/deepseek_mla_decode.py
@@ -6,7 +6,7 @@ from tileops.kernels.attention import MlaDecodeKernel, MlaDecodeWsKernel
 from tileops.kernels.kernel_base import Kernel
 from tileops.utils import is_hopper
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["MultiHeadLatentAttentionDecodeWithKVCacheFwdOp"]
