+++ tileops/ops/attention/deepseek_nsa.py
@@ -10,7 +10,7 @@ from tileops.kernels.attention import (
 )
 from tileops.kernels.kernel_base import Kernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = [
     "MeanPoolingForwardOp",
