+++ tileops/ops/attention/gqa_decode_paged.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.attention.gqa_decode_paged import GqaDecodePagedKernel
 from tileops.kernels.kernel_base import Kernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["GroupedQueryAttentionDecodePagedWithKVCacheFwdOp"]
