+++ tileops/ops/attention/mha.py
@@ -13,7 +13,7 @@ from tileops.kernels.attention import (
 from tileops.kernels.kernel_base import Kernel
 from tileops.utils import is_hopper
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["MultiHeadAttentionFwdOp", "MultiHeadAttentionBwdOp"]
