+++ tileops/ops/attention/mha_decode_paged.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.attention.mha_decode_paged import MhaDecodePagedKernel
 from tileops.kernels.kernel_base import Kernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["MultiHeadAttentionDecodePagedWithKVCacheFwdOp"]
