+++ tileops/ops/attention/mhc_post.py
@@ -5,7 +5,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.mhc import MHCPostKernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["MHCPostOp"]
