+++ tileops/ops/moe/fused_topk.py
@@ -7,7 +7,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.moe.fused_topk import FusedTopKKernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["FusedTopKOp"]
