+++ tileops/ops/moe/permute_nopad.py
@@ -7,7 +7,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.moe.permute_nopad import MoePermuteNopadKernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["MoePermuteNopadFwdOp"]
