+++ tileops/ops/norm/base.py
@@ -14,7 +14,7 @@ import torch.nn.functional as F
 
 from tileops.kernels.kernel_base import Kernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["RowNormOp", "ALIGNMENT"]
