+++ tileops/ops/norm/layer_norm.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.norm import LayerNormKernel
 
-from ..op import Op
+from ..op_base import Op
 
 __all__ = ["LayerNormFwdOp"]
