+++ tileops/ops/reduction/_softmax_base.py
@@ -17,7 +17,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.reduction._primitives import DEFAULT_ALIGNMENT, align_up
 
-from ..op import Op
+from ..op_base import Op
 from ._multidim import flatten_for_multidim, normalize_dim, restore_multidim_shape
 
 __all__ = ["_SoftmaxBaseOp"]
