+++ tileops/ops/reduction/reduce.py
@@ -22,7 +22,7 @@ from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.reduction._primitives import DEFAULT_ALIGNMENT, align_up
 from tileops.kernels.reduction.reduce import ReduceKernel
 
-from ..op import Op
+from ..op_base import Op
 from ._multidim import flatten_for_multidim, normalize_dim, restore_multidim_shape
 
 __all__ = [
